--- Project3.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32825.248
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Project3", "Project3\Project3.vcxproj", "{7EC8898D-9272-4049-8094-121A3F1649B3}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Debug|x64.ActiveCfg = Debug|x64
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Debug|x64.Build.0 = Debug|x64
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Debug|x86.ActiveCfg = Debug|Win32
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Debug|x86.Build.0 = Debug|Win32
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Release|x64.ActiveCfg = Release|x64
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Release|x64.Build.0 = Release|x64
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Release|x86.ActiveCfg = Release|Win32
-		{7EC8898D-9272-4049-8094-121A3F1649B3}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E806E34F-0870-4638-B1E6-D14AC28459C1}
-	EndGlobalSection
-EndGlobal
