--- Project3/Address.cpp
@@ -1,4 +1,0 @@
-#include "Address.h"
-void Address::display() const {
-    cout << "Адреса: " << city << ", " << street << " (ZIP: " << zipCode << ")" << endl;
-}
