--- Project3/RecordBook.cpp
@@ -1,10 +1,0 @@
-#include "RecordBook.h"
-void RecordBook::addSubject(Subject s) {
-    subjects.push_back(s);
-}
-void RecordBook::display() const {
-    cout << "Залікова книжка №: " << recordNumber << endl;
-    for (const auto& sub : subjects) {
-        sub.display();
-    }
-}
