--- Project3/Student.cpp
@@ -1,8 +1,0 @@
-#include "Student.h"
-Student::Student(string n, Address a, string g, RecordBook rb) : name(n), address(a), group(g), recordBook(rb) {}
-void Student::display() const {
-    cout << "Студент: " << name << endl;
-    address.display();
-    cout << "Група: " << group << endl;
-    recordBook.display();
-}
