--- Project3/Subject.cpp
@@ -1,5 +1,0 @@
-#include "Subject.h"
-void Subject::display() const {
-    cout << "Предмет: " << name << " | Семестр: " << semester << " | Оцінка: " << grade << endl;
-}
-
