--- Project3/Subject.h
@@ -1,15 +1,0 @@
-#ifndef SUBJECT_H
-#define SUBJECT_H
-#include <iostream>
-using namespace std;
-
-class Subject {
-public:
-    string name;
-    int semester;
-    double grade;
-
-    Subject(string n, int sem, double g) : name(n), semester(sem), grade(g) {}
-    void display() const;
-};
-#endif
