--- Project3/x64/Debug/Project3.log
@@ -1,2 +1,0 @@
-﻿  main.cpp
-  Project3.vcxproj -> C:\Users\User\source\repos\Project3\x64\Debug\Project3.exe
