--- Project3/x64/Debug/Project3.tlog/Project3.lastbuildstate
@@ -1,2 +1,0 @@
-PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.33.31629:TargetPlatformVersion=10.0.19041.0:
-Debug|x64|C:\Users\User\source\repos\Project3\|
