+++ lab01/InitData.h
@@ -1,0 +1,46 @@
+#ifndef INITDATA_H
+#define INITDATA_H
+
+#include "Student.h"
+#include <string>
+
+void initializeStudents(Student students[], int& count) {
+    string names[] = {
+        "Анастасія Купчак", "Олег Бойко", "Марина Іваненко", "Андрій Петренко",
+        "Сергій Коваленко", "Олександр Грицай", "Юлія Мельник", "Іван Сидоренко",
+        "Вікторія Ткаченко", "Дмитро Романов"
+    };
+
+    Address addresses[] = {
+        Address("79000", "Львів", "Шевченка 10"),
+        Address("01001", "Київ", "Хрещатик 1"),
+        Address("65000", "Одеса", "Дерибасівська 5"),
+        Address("46000", "Тернопіль", "Сагайдачного 12"),
+        Address("58000", "Чернівці", "Головна 25"),
+        Address("76000", "Івано-Франківськ", "Незалежності 8"),
+        Address("79001", "Львів", "Городоцька 90"),
+        Address("10001", "Житомир", "Михайлівська 3"),
+        Address("50000", "Кривий Ріг", "Пушкіна 14"),
+        Address("40000", "Суми", "Харківська 20")
+    };
+
+    string groups[] = { "КН-2023", "ПМ-2022", "ФІ-2021", "МЕ-2023", "БІ-2022",
+                       "КН-2023", "ПМ-2021", "ФІ-2022", "МЕ-2021", "БІ-2023" };
+
+    string subjects[] = { "Математика", "Програмування", "Фізика" };
+    int grades[][3] = {
+        {85, 90, 80}, {78, 82, 85}, {88, 75, 92}, {91, 85, 89},
+        {70, 65, 80}, {95, 88, 92}, {60, 78, 85}, {89, 85, 90},
+        {77, 92, 79}, {80, 75, 88}
+    };
+
+    for (int i = 0; i < 10; i++) {
+        RecordBook rb(to_string(123450 + i));
+        for (int j = 0; j < 3; j++) {
+            rb.addSubject(Subject(subjects[j], j + 1, grades[i][j]));
+        }
+        students[count++] = Student(names[i], addresses[i], groups[i], rb);
+    }
+}
+
+#endif
